userDetail 관련 기능 추가(editUser, changePassword, protect route)

+++ controllers/videoController.js
@@ -40,7 +40,10 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -49,7 +52,7 @@ export const videoDetail = async (req, res) => {
     const {
       params: { id },
     } = req;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageName: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -61,8 +64,13 @@ export const getEditVideo = async (req, res) => {
     const {
       params: { id },
     } = req;
-    const video = await Video.findById(id);
-    res.render("editVideo", { pageName: `Edit ${video.title}`, video });
+    const video = await Video.findById(id).populate("creator");
+
+    if (video.creator.id !== req.user.id) {
+      throw Error;
+    } else {
+      res.render("editVideo", { pageName: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -86,7 +94,12 @@ export const deleteVideo = async (req, res) => {
     const {
       params: { id },
     } = req;
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error;
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
